Add: set quantity setter to float

+++ src/main/java/com/tank/tankadminbackend/models/unas/order/Item.java
@@ -39,7 +39,7 @@ public class Item {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(float quantity) {
         this.quantity = quantity;
     }
